app/requirement.txt
Flask
Werkzeug
joblib
pandas
numpy
matplotlib
scikit-learn
tensorflow
Pillow
ssl